+++ adagucserverEC/utils/CRequestUtils.h
@@ -23,5 +23,11 @@ std::vector<std::string> getReferenceTimes(CDataSource &dataSource);
 
 int getMaxQueryLimit(CDataSource &dataSource);
 
+/**
+ * Time formatting / Formats time to correct iso8601 string with a 'T' at position 10 and a trailing 'Z'.
+ * If the timestring is shorter then the same string is simply returned.
+ * @param input time string like '2025-01-02 01:02:03`
+ * @returns formatted iso string like  '2025-01-02T01:02:03Z`
+ */
 std::string makeIsoStringFromDbString(std::string input);
 #endif