+++ hclasses/CDirReader.cpp
@@ -35,12 +35,14 @@
 #include <algorithm> /* For std::sort */
 #include <regex>
 #include "CDirReader.h"
+#include "CTString.h"
+#include "CDebugger.h"
 
 CDirReader::CDirReader() {}
 
 CDirReader::~CDirReader() { fileList.clear(); }
 
-std::vector<std::string> CDirReader::listDir(const char *directory, bool recursive, const char *ext_filter, int filesAndOrDirs, bool exceptionOnError) {
+const std::vector<std::string> CDirReader::listDir(const char *directory, bool recursive, const char *ext_filter, int filesAndOrDirs, bool exceptionOnError) {
   std::vector<std::string> result;
   std::regex self_regex;
   if (ext_filter != nullptr) {
@@ -52,9 +54,9 @@ std::vector<std::string> CDirReader::listDir(const char *directory, bool recursi
   if ((dir = opendir(directory)) != NULL) {
     /* print all the files and directories within directory */
     while ((ent = readdir(dir)) != NULL) {
-      CT::string fullName = directory;
-      fullName.concat("/");
-      fullName.concat(ent->d_name);
+      std::string fullName = directory;
+      fullName += fullName + "/" + ent->d_name;
+
       // Deal with filesystems that don't provide d_type
       auto d_type = ent->d_type;
       if (d_type == DT_UNKNOWN) {
@@ -77,8 +79,7 @@ std::vector<std::string> CDirReader::listDir(const char *directory, bool recursi
           }
           if (recursive && d_type == DT_DIR) {
             auto dirFiles = listDir(fullName.c_str(), recursive, ext_filter, filesAndOrDirs, exceptionOnError);
-            // Move elements from dirFiles to result.
-            // dirFiles is left in undefined but safe-to-destruct state.
+            // Move elements from dirFiles to result. dirFiles is left in undefined but safe-to-destruct state.
             result.insert(result.end(), std::make_move_iterator(dirFiles.begin()), std::make_move_iterator(dirFiles.end()));
             dirFiles.clear();
           }
@@ -93,21 +94,16 @@ std::vector<std::string> CDirReader::listDir(const char *directory, bool recursi
       throw "Error";
     }
   }
-
-  // if (re != nullptr && ext_filter != nullptr) {
-  //   regfree(re);
-  // }
   return result;
 }
 
 int CDirReader::listDirRecursive(const char *directory, const char *ext_filter) {
   if (fileList.size() != 0) {
-    if (!currentDir.equals(directory)) {
+    if (currentDir != directory) {
       CDBError("Trying to do listDirRecursive twice with different paths");
       return 1;
     } else {
       /* Already done */
-      // CDBDebug("Using cached results for [%s]", directory);
       return 0;
     }
   }
@@ -132,7 +128,7 @@ int CDirReader::_listDirRecursive(const char *directory, const char *ext_filter)
       return 2;
     }
     /* Check filter*/
-    if (testRegEx(directory, ext_filter) == 1) {
+    if (CT::testRegEx(directory, ext_filter)) {
       fileList.push_back(makeCleanPath(directory).c_str());
     } else {
       CDBWarning("Regexp failed.");
@@ -142,45 +138,29 @@ int CDirReader::_listDirRecursive(const char *directory, const char *ext_filter)
   }
 }
 int CDirReader::_ReadDir(const char *directory, const char *ext_filter, int recursive) {
-  DIR *dp;
   struct dirent *ep;
-  int filename_len;
-  int dir_len = strlen(directory);
-
-  dp = opendir(directory);
-  if (dp != NULL) {
-    ep = readdir(dp);
-    while (ep) {
-      filename_len = strlen(ep->d_name);
-      CT::string dirpath;
-      dirpath.copy(directory, dir_len);
-      dirpath.concatlength("/", 1);
-      dirpath.concatlength(ep->d_name, filename_len);
-      bool isdir = isDir(dirpath.c_str());
-      if (isdir == false) {
-        if (testRegEx(ep->d_name, ext_filter) == 1) {
-
-          /* This is a file */
-          CT::string fullName;
-          filename_len = strlen(ep->d_name);
-          fullName.copy(directory, dir_len);
-          fullName.concatlength("/", 1);
-          fullName.concatlength(ep->d_name, filename_len);
-          fullName = makeCleanPath(fullName.c_str());
-          fileList.push_back(fullName.c_str());
-        }
-      } else if (recursive == 1) {
-        if (ep->d_name[0] != '.') {
-          _ReadDir(dirpath.c_str(), ext_filter, recursive);
-        }
-      }
-      ep = readdir(dp);
-    }
-    (void)closedir(dp);
-  } else {
-    /* Could not open the directory */
+  auto dp = opendir(directory);
+  if (dp == NULL) {
     throw 1;
   }
+  ep = readdir(dp);
+  while (ep) {
+    std::string fullName = std::string(directory) + "/" + ep->d_name;
+    bool isdir = isDir(fullName.c_str());
+    if (isdir == false) {
+      if (CT::testRegEx(ep->d_name, ext_filter)) {
+        /* This is a file */
+        fileList.push_back(makeCleanPath(fullName));
+      }
+    } else if (recursive == 1) {
+      if (ep->d_name[0] != '.') {
+        _ReadDir(fullName.c_str(), ext_filter, recursive);
+      }
+    }
+    ep = readdir(dp);
+  }
+  (void)closedir(dp);
+
   return 0;
 }
 
@@ -200,34 +180,17 @@ bool CDirReader::isFile(const char *filename) {
   return S_ISREG(fileattr.st_mode);
 }
 
-int CDirReader::testRegEx(const char *string, const char *pattern) {
-  if (string == nullptr || pattern == nullptr) return 1;
-  int status;
-  regex_t re;
-
-  if (regcomp(&re, pattern, REG_EXTENDED | REG_NOSUB) != 0) {
-    return (0); /* Report error. */
-  }
-  status = regexec(&re, string, (size_t)0, NULL, 0);
-  regfree(&re);
-  if (status != 0) {
-    return (0); /* Report error. */
-  }
-
-  return (1);
-}
-
 void CDirReader::makePublicDirectory(const char *dirname) {
   if (dirname == NULL) return;
   struct stat stFileInfo;
   int intStat = stat(dirname, &stFileInfo);
   if (intStat != 0) {
-    CT::string directory = dirname;
-    auto directorySplitted = directory.split("/");
+    std::string directory = dirname;
+    auto directorySplitted = CT::split(directory, "/");
     directory = "";
     for (auto &directoryPart: directorySplitted) {
-      directory.concat("/");
-      directory.concat(directoryPart);
+      directory += "/";
+      directory += directoryPart;
       const char *part = directory.c_str();
       int intStat = stat(part, &stFileInfo);
       if (intStat != 0) {
@@ -290,7 +253,7 @@ void CDirReader::test_compareLists() {
 }
 
 int CDirReader::test_makeCleanPath() {
-  CT::string t;
+  std::string t;
   t = makeCleanPath("data");
   CDBDebug("[%s]", t.c_str());
   ;
@@ -390,8 +353,7 @@ std::string getFileDate(const std::string &fileName) {
 
   std::map<std::string, std::string>::iterator it = lookupTableFileModificationDateMap.find(fileName);
   if (it != lookupTableFileModificationDateMap.end()) {
-    CT::string filemoddate = (*it).second.c_str();
-    return filemoddate;
+    return it->second;
   }
   auto fileDate = getFileDateUnCached(fileName);
   if (fileDate.length() < 10) fileDate = ("1970-01-01T00:00:00Z");